+++ REQUIRE
@@ -3,6 +3,4 @@ GZip 0.2.18
 LightXML 0.2.1
 ParserCombinator 1.7.3
 JLD
-Clustering 0.4
 Distributions
-MatrixDepot
