--- src/datasets/Datasets.jl
@@ -1,14 +1,0 @@
-module Datasets
-
-using ...LightGraphs
-using MatrixDepot
-# smallgraphs
-export smallgraph,
-
-# matrixdepot
-MDGraph, MDDiGraph
-
-include("smallgraphs.jl")
-include("matrixdepot.jl")
-
-end
