--- test/datasets/runtests.jl
@@ -1,11 +1,0 @@
-tests = [
-    "datasets/smallgraphs",
-    "datasets/matrixdepot"
-]
-
-
-for t in tests
-    tp = joinpath(testdir,"$(t).jl")
-    println("running $(tp) ...")
-    include(tp)
-end
